requirements.txt

fastapi>=0.116.1
uvicorn[standard]>=0.35.0
pydantic>=2.11.7
pydantic-settings>=2.10.1
python-jose[cryptography]>=3.5.0
APScheduler>=3.11.0
python-dotenv>=1.1.1
SQLAlchemy>=2.0.43
psycopg2-binary>=2.9.10
twilio>=9.7.0
tenacity>=9.1.2
prometheus-client>=0.22.1
